requirements/base.txt
aldjemy==0.3.10
defusedxml==0.5.0
Django==1.11.3
django-auth-ldap==1.2.16
django-extensions==2.1.2
django-mathfilters==0.4.0
django-simple-history==1.9.0
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
dnspython==1.15.0
email-validator==1.0.2
env-config==1.10.0
fake-factory==9999.9.9
Faker==0.8.3
graphviz==0.9
idna==2.6
ipaddress==1.0.18
pathlib==1.0.1
psycopg2==2.7.1
pydot==1.2.4
pyparsing==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
python-ldap==2.4.44
pytz==2017.2
selenium==3.141.0
six==1.11.0
SQLAlchemy==1.1.11
typing==3.6.6
urllib3==1.24.1
